+++ src/Session.php
@@ -14,7 +14,6 @@ class Session implements SessionInterface
     protected $accessed;
     protected $created;
     protected $updated;
-    protected $expires;
     protected $locked = false;
 
     public function __construct(
@@ -22,22 +21,19 @@ class Session implements SessionInterface
         array $data,
         DateTimeInterface $created,
         DateTimeInterface $accessed,
-        DateTimeInterface $updated,
-        DateTimeInterface $expires
+        DateTimeInterface $updated
     ) {
         $this->id = $id;
         $this->data = $data;
         $this->created = new DateTimeImmutable($created->format('YmdHis'), $created->getTimezone());
         $this->accessed = new DateTimeImmutable($accessed->format('YmdHis'), $accessed->getTimezone());
         $this->updated = new DateTimeImmutable($updated->format('YmdHis'), $updated->getTimezone());
-        $this->expires = new DateTimeImmutable($expires->format('YmdHis'), $expires->getTimezone());
     }
 
-    public static function createWithId(SessionId $id, $interval = 'PT3M')
+    public static function createWithId(SessionId $id)
     {
         $accessed = $created = $updated = new DateTimeImmutable('now', new DateTimeZone('UTC'));
-        $expires = $updated->add(new DateInterval($interval));
-        $session = new static($id, [], $created, $accessed, $updated, $expires);
+        $session = new static($id, [], $created, $accessed, $updated);
         return $session;
     }
 
@@ -86,26 +82,6 @@ class Session implements SessionInterface
     public function getUpdated(): DateTimeImmutable
     {
         return $this->updated;
-    }
-
-    public function getExpires(): DateTimeImmutable
-    {
-        return $this->expires;
-    }
-
-    public function renew($interval = 'PT3M')
-    {
-        $this->markAsUpdated();
-        $this->expires = $this->updated->add(new DateInterval($interval));
-    }
-
-    public function isExpired(DateTimeInterface $when = null): bool
-    {
-        if (is_null($when)) {
-            $when = new DateTimeImmutable('now', new DateTimeZone('UTC'));
-        }
-
-        return ($this->expires <= $when);
     }
 
     public function lock()