+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this package will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## 0.3.0 - 2017-03-04
+### Changed
+- Switched back to __get, __set, __isset, and __unset magic methods from get, set, has, and remove methods
+- Moved expired tracking from Session to Storage
+### Removed
+- Renew, getExpires, and isExpired from Session and SessionInterface
+
 ## 0.2.1 - 2017-02-06
 ### Added
 - Finishing a session locks the DomainSession object from further updates